Added a .devcontainer containing instructions for setting up a development container for Ceph on a mac
This method makes setting up Ceph much easier on a mac.
Rather than using docker, it uses other forms of technology
which is more efficient than docker and have all
the necessary to easily set up Ceph. Additionally,
I made a shell script that immediately sets up
the environment for Ceph. It can be used for future
RCOS or other students that want to work on ceph and have
a mac with sufficient storage.

Signed-off-by: Jordan Ye <yejordan8888@gmail.com>