Fixed errors in the documentation.
Realized that there are significant errors in the documentation.
Additionally, there are some better ways and more efficient ways to do
run the commands which I likely will need to utilize
in order to save time and not have to redo the work every session.
Perhaps, enviorment variables can be used to store the values
and shell script to automatically start up the process of setting up the
Ceph cluster with the keys I need for testing.

Signed-off-by: Jordan Ye <yejordan8888@gmail.com>